refactor(consumer): extract KPipeConsumer.Builder to a top-level KPipeConsumerBuilder
KPipeConsumer was 1974 lines — a god-class whose ~570-line nested Builder
dominated it and hurt navigability. Extract the Builder to its own top-level
KPipeConsumerBuilder.java in the same package (§16 delete-and-migrate).

- KPipeConsumer.java: 1974 → 1404 lines.
- New public final class KPipeConsumerBuilder (package-private ctor, so
  KPipeConsumer.builder() is still the only entry point — the original nested
  Builder had a private ctor, so `final` faithfully preserves "not
  subclassable" and makes it explicit).
- The 28 config fields the KPipeConsumer ctor consumes are widened private →
  package-private (the exact minimal set the compiler demands — the builder's
  fields ARE the config the product reads). Effective access is unchanged: the
  nested Builder already exposed these to KPipeConsumer via shared-private
  access; package-private keeps the same two effective readers within one
  package, with no external access.
- BatchSpec<T> moves with the builder (KPipeConsumerBuilder.BatchSpec).
- builder() now returns KPipeConsumerBuilder; the public ctor param retyped.
- Migrated the 3 explicit-type references (DefaultStream, MultiBuilder,
  KPipeConsumerBuilderValidationTest) + 8 javadoc/comment refs. The ~85
  `var builder = KPipeConsumer.builder()` call sites need no change.

Verified: :lib:kpipe-consumer:test green (full jcstress/property/chaos suite,
3m18s); consumer + api + api-test compile green; zero remaining
`KPipeConsumer.Builder` references.
