+++ lib/kpipe-api/src/main/java/io/github/eschizoid/kpipe/DefaultBatchSink.java
@@ -10,7 +10,7 @@ import java.util.Objects;
 /// Package-private [Sink] impl for batch terminals. Builds a sink-less
 /// [io.github.eschizoid.kpipe.registry.MessagePipeline]
 /// from the stream's operator chain (deserialize → operators → return value) and wires it into
-/// [KPipeConsumer.Builder#withBatchPipeline] together with the configured [BatchSink] and
+/// [KPipeConsumerBuilder#withBatchPipeline] together with the configured [BatchSink] and
 /// [BatchPolicy]. Honors whatever processing mode the underlying [DefaultStream] carries
 /// (default: parallel); consumer-level config (metrics, errorHandler, DLQ, pollTimeout, retry,
 /// backpressure, tracer, circuit breaker) carries over too via
