+++ lib/kpipe-api/src/main/java/io/github/eschizoid/kpipe/Sink.java
@@ -20,7 +20,7 @@ public interface Sink<T> {
   /// Starts the pipeline. The returned [Handle] is the only way to query metrics, check
   /// health, or shut down the running consumer. The handle is `AutoCloseable` — wrap the call in
   /// try-with-resources to guarantee shutdown. For long-running hosts, opt in to a JVM shutdown
-  /// hook via `KPipeConsumer.Builder.withShutdownHook(true)` on the explicit-API path.
+  /// hook via `KPipeConsumerBuilder.withShutdownHook(true)` on the explicit-API path.
   ///
   /// @return a [Handle] for monitoring and shutdown
   Handle start();
