+++ lib/kpipe-consumer/src/test/java/io/github/eschizoid/kpipe/consumer/KPipeDlqTest.java
@@ -197,7 +197,7 @@ class KPipeDlqTest {
     }
   }
 
-  /// Exercises the atomic [KPipeConsumer.Builder#withDeadLetterQueue(String, KPipeProducer)]
+  /// Exercises the atomic [KPipeConsumerBuilder#withDeadLetterQueue(String, KPipeProducer)]
   /// setter: a single call pairs the DLQ topic with a pre-built [KPipeProducer], and the consumer
   /// routes failures through that producer to that topic. The two-call alternative
   /// (`withDeadLetterTopic` + `withKafkaProducer`) is covered by the tests above; this one
